Hangman Game
A secret word  is selected at random, and users are shown the number of letters as blank lines. Users guess letters one at a time. If the secret word contains the guessed letter, all instances of that letter are revealed. If the letter is incorrect, the player loses one of their 6 guesses. If all letters are revealed, the player wins. If the player loses 
their last guess, they lose and the answer should be revealed.